Stop displaying a zero length unallocated space after every LV (#316)
A zero length unallocated Partition object was displayed after every
Logical Volume.  This was because insert_unallocated() used integer
division of MEBIBYTE / sector_size to check for gaps.  But for Volume
Groups the sector_size is set to the PE (Physical Extent) size which is
4 MiB.  MEBIBYTE / 4194304 = 0, so a gap >= 0 resulted in a spurious
unallocated Partition object being inserted.

Fix this by computing and comparing the gap in units of bytes.

Closes #316 - Add read-only LVM Volume Group support