Line wrap file system status value in the Information dialog (#771693)
Again this is to handle a long list of mount points for a single
partition.  To prevent the file system status value of "Mounted on [list
of 20 mount points]" causing the dialog to become wider than the screen,
line wrap the text.  This instead makes the dialog taller, which already
automatically scrolls vertically as needed.

Bug 771693 - Mount Point column is wider than the screen on openSUSE
             using btrfs root with lots of mounted subvolumes